requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Bcrypt
python-dotenv
Flask-WTF
WTForms
SQLAlchemy
